[feat] path refactor
