+++ project/src/main/java/ict/project/resume/entity/ResumeEntity.java
@@ -1,5 +1,6 @@
 package ict.project.resume.entity;
 
+import ict.project.user.UserEntity;
 import jakarta.persistence.*;
 import lombok.*;
 import java.time.LocalDateTime;
